Added Kali Setup script
This script automates my personal setup of Kali Linux. It installs some tools that aren't included in the base Kali installer, sets up some folder structure that I use for services like TryHackMe and HackTheBox. I've identified some things that I could improve so a cleaner version of this will be in the works but I needed to get this in a repo quickly.